refactor(coding): separate CLI entrypoint from application bindings
