+++ docs/internals/architecture/harness/plugin/plugin-lifecycle-plc9c5-c50-inventory.md
@@ -26,7 +26,7 @@ transfer authority to Product or Plugin management.
 | --- | --- | --- | --- |
 | `C5-CUR-DECL` | `src/loushang/harness/resources/plugins/declarations.py` and `src/loushang/harness/resources/plugins/selection.py` | versioned inert `local_worker` declaration/selection exists for `capability_provider`; declared required/optional and exact Worker configuration are data only | retain; Product must rejoin the exact selected reservation and immutable revision before a receipt and cannot downgrade declared requiredness |
 | `C5-CUR-PRODUCT-CORE` | `src/loushang/harness/session/bootstrap_construction.py` and `src/loushang/harness/session/agent_product.py` | one Product-bound Agent Session construction path exists; it owns Product callbacks and final Session assembly but has no Worker activation input | reuse as the shared canary-capable construction boundary; do not put activation in presenters |
-| `C5-CUR-CODING-PRODUCT` | `src/loushang/coding/bootstrap.py`, `src/loushang/coding/cli/__main__.py`, and `src/loushang/coding/_product_worker_canary.py` | Coding supplies an explicit `product_id`, converges ordinary Agent/TUI/RPC/channel/plain/workflow modes through one runtime builder, and owns the sole explicit Linux/Windows Worker canary composition; early-dispatch workspace/LSP/multiagent routes remain separate | implemented C5.4/C5.5c Product evidence; the one canary consumes the public Harness Worker facade plus the exact reviewed coordinator/friend seams and keeps every excluded route dark |
+| `C5-CUR-CODING-PRODUCT` | `src/loushang/coding/bootstrap.py`, `src/loushang/coding/cli/application.py`, and `src/loushang/coding/_product_worker_canary.py` | Coding supplies an explicit `product_id`, converges ordinary Agent/TUI/RPC/channel/plain/workflow modes through one runtime builder, and owns the sole explicit Linux/Windows Worker canary composition; early-dispatch workspace/LSP/multiagent routes remain separate | implemented C5.4/C5.5c Product evidence; the one canary consumes the public Harness Worker facade plus the exact reviewed coordinator/friend seams and keeps every excluded route dark |
 | `C5-CUR-PACKAGE-PRODUCT` | `src/loushang/harness/resources/packages/product_contract.py`, `src/loushang/harness/resources/packages/product_runtime.py`, `src/loushang/harness/resources/packages/product_activation.py`, and `src/loushang/harness/resources/packages/product_composition.py` | PLC9A2 owns Package lifecycle routing for CLI/RPC/Session/startup/operations; it is not a generic Product catalog and contains no Worker authority | retain as Product composition precedent; do not extend its Package-specific receipt into Worker activation |
 | `C5-CUR-SESSION-PROFILE` | `src/loushang/harness/transcript/runtime_profile.py` and `src/loushang/coding/session_manager.py` | Coding persists a Product-bound runtime profile and refuses foreign Product profile snapshots on restore | reuse for the Coding-specific G7 resume row after locator selection; do not claim generic AppHost pre-routing identity |
 | `C5-CUR-SESSION-DISCOVERY` | `src/loushang/harness/transcript/discovery.py`, `src/loushang/harness/transcript/directory.py`, `src/loushang/harness/transcript/session_catalog.py`, and `src/loushang/harness/machine_resources/control_plane.py` | canonical global plus cwd/home compatibility sources have stable source/locator, alias/conflict, and read-only projection semantics | retain; discovery selects a candidate and grants no Worker/Product authority |
