+++ tests/architecture/test_foreground_hosted_application_g12.py
@@ -150,6 +150,7 @@ def test_G12_EXPLICIT_ACTIVATION_preserves_installed_omission_routes() -> None:
         Path("src/loushang/coding/__init__.py"),
         Path("src/loushang/coding/bootstrap.py"),
         Path("src/loushang/coding/cli/__main__.py"),
+        Path("src/loushang/coding/cli/application.py"),
         Path("src/loushang/coding/ui/cli.py"),
         Path("src/loushang/harnesstui/__init__.py"),
         Path("src/loushang/harnesstui/conversation/application_host.py"),
