+++ tests/architecture/test_foreground_stdio_hosted_app_g14.py
@@ -105,7 +105,7 @@ def test_G14_BOUNDARIES_only_explicit_product_command_composes_native_stdio() ->
     assert "loushang.appserver.local" not in bootstrap
     assert "loushang.appserver.stdio" not in bootstrap
     for default in (
-        "src/loushang/coding/cli/__main__.py",
+        "src/loushang/coding/cli/__main__.py", "src/loushang/coding/cli/application.py",
         "src/loushang/coding/ui/cli.py",
         "src/loushang/coding/__init__.py",
     ):
